backend/ai-requirements.txt
absl-py==0.12.0
asgiref==3.3.1
astunparse==1.6.3
boto3==1.17.39
botocore==1.20.39
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.1.7
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0

easydict==1.9
flatbuffers==1.12
gast==0.3.3
google-api-core==1.26.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-cloud-speech==2.2.0
google-cloud-texttospeech==2.2.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libcst==0.3.17
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.0
mypy-extensions==0.4.3

numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
Pillow==8.1.2
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
# PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.0
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
scipy==1.6.2
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
